perf(test): move shared helpers out of test files
Test files importing helpers from other test files caused vitest to
re-register the imported file's top-level describe blocks. Tests like
`AuthenticationService > Hash Authentication > ...` were running in 5
test files; `InvoiceService > createInvoice > ...` in 3 test files.

Move `userIsAsExpected` to test/helpers/authentication-helpers.ts and
`createInvoiceWithTransfers` to test/helpers/invoice-helpers.ts. The
helpers directory is excluded from the vitest include glob, so these
no longer pull along describe blocks.
