+++ test/helpers/authentication-helpers.ts
@@ -1,0 +1,33 @@
+/**
+ *  SudoSOS back-end API service.
+ *  Copyright (C) 2026 Study association GEWIS
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU Affero General Public License as published
+ *  by the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU Affero General Public License for more details.
+ *
+ *  You should have received a copy of the GNU Affero General Public License
+ *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ *
+ *  @license
+ */
+
+import { expect } from 'chai';
+import User from '../../src/entity/user/user';
+import { UserResponse } from '../../src/controller/response/user-response';
+import { isNumber } from '../../src/helpers/validators';
+
+export default function userIsAsExpected(user: User | UserResponse, ADResponse: any) {
+  expect(user.firstName).to.equal(ADResponse.givenName);
+  expect(user.lastName).to.equal(ADResponse.sn);
+  if (isNumber(user.type)) expect(user.type).to.equal(1);
+  expect(user.active).to.equal(true);
+  expect(user.deleted).to.equal(false);
+  expect(user.canGoIntoDebt).to.equal(true);
+}
