+++ test/unit/gewis/gewis.ts
@@ -30,7 +30,7 @@ import User, { UserType } from '../../../src/entity/user/user';
 import Database from '../../../src/database/database';
 import seedDatabase from '../../seed';
 import Swagger from '../../../src/start/swagger';
-import userIsAsExpected from '../service/authentication-service';
+import userIsAsExpected from '../../helpers/authentication-helpers';
 import { inUserContext, UserFactory } from '../../helpers/user-factory';
 import MemberUser from '../../../src/entity/user/member-user';
 import Gewis from '../../../src/gewis/gewis';
